requirements.txt
streamlit
pandas
sqlalchemy
groq
python-dotenv
matplotlib
plotly
openpyxl
numpy
anthropic